Simplify API endpoints
Minor code refactor
