--- kinvo/api/routes/heatbeat.py
@@ -1,9 +1,0 @@
-from flask import Blueprint
-
-
-heartbp = Blueprint("heartbp", __name__)
-
-
-@heartbp.route("/alive", methods=['GET'])
-def alive():
-    return {'is_alive': True}, 200
