+++ kinvo/main.py
@@ -1,6 +1,8 @@
+import spacy
+
 from flask import Flask
 
-from kinvo.api.router import register_routes
+from kinvo.api import routesbp
 from kinvo.core import config
 from kinvo.logger import logger
 
@@ -14,14 +16,16 @@ def create_app():
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
     app.config['SECRET_KEY'] = config.SECRET_KEY
 
+    app.nlp_model = spacy.load("pt_core_news_sm")
+
     # Initialize API routers
-    register_routes(app)
+    app.register_blueprint(routesbp, url_prefix=config.API_PREFIX)
 
     # Define a hello world page
-    @app.route('/')
+    @ app.route('/')
     def hello_world():
         return '<h1>Hello, World!</h1>'
 
-    logger.info(f'App is up!')
+    logger.info(f'App is ready!')
 
     return app
